requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==20.1.0
protobuf==4.21.7
psycopg2==2.9.4
pybind11==2.10.0
pycairo==1.21.0
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pytz==2022.4
SciPy==1.9.1
sqlparse==0.4.3
whitenoise==6.2.0
